I noticed that monitor encryption is not working on 64bit plattforms due to other sizes of long on this plattform. This patch converts all numeric types like long to portable types like uint32_t. This is not complete yet (only base folder up to r4987, changes thereafter might be incompletely converted). If you encounter problems through this patch please report in forum.
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@4994 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ csctapi/ifd_phoenix.c
@@ -211,7 +211,7 @@ int Phoenix_Receive (struct s_reader * reader, BYTE * buffer, unsigned size, uns
 	return OK;
 }
 
-int Phoenix_SetBaudrate (struct s_reader * reader, unsigned long baudrate)
+int Phoenix_SetBaudrate (struct s_reader * reader, uint32_t baudrate)
 {
 	cs_debug_mask (D_IFD, "IFD: Phoenix Setting baudrate to %lu\n", baudrate);
